LICENSE Removed useless docs

+++ simElectricity/src/main/java/simElectricity/Common/EnergyNet/MatrixOperation.java
@@ -1,3 +1,22 @@
+/*
+ * Copyright (C) 2014 SimElectricity
+ *
+ * This library is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU Lesser General Public
+ * License as published by the Free Software Foundation; either
+ * version 2.1 of the License, or (at your option) any later version.
+ *
+ * This library is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ * Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public
+ * License along with this library; if not, write to the Free Software
+ * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301
+ * USA
+ */
+
 package simElectricity.Common.EnergyNet;
 
 public class MatrixOperation {
@@ -44,7 +63,7 @@ public class MatrixOperation {
         float[] x = new float[N];
         for (int i = N - 1; i >= 0; i--) {
             if (A[i][i] != 0) {//Ignore any line with all zero
-                double sum = (double) 0.0;
+                double sum = 0.0;
                 for (int j = i + 1; j < N; j++) {
                     sum += A[i][j] * x[j];
                 }
